More Service

+++ src/main/java/com/example/cineket/service/IPeliculaService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IPeliculaService {
 
     List<Pelicula> getPeliculas();
-    Pelicula getPeliculaById(int id);
+    Pelicula getPeliculaById(Long id);
     Pelicula createPelicula(Pelicula pelicula);
     Pelicula updatePelicula(Long id, Pelicula pelicula);
     void deletePelicula(Long id);
